Move Dokploy target setup to FastAPI
