+++ docs/service-boundary.md
@@ -160,7 +160,9 @@ VeriReel product paths:
 - product onboarding route:
   - `POST /v1/product-onboarding/apply`
 - Dokploy target setup route:
-  - `POST /v1/dokploy-targets/setup`
+  - `POST /v1/dokploy-targets/setup` (native FastAPI for bearer-token
+    callers, DB-backed setup records, apply-only `Idempotency-Key`
+    replay/conflict handling, and repeatable dry-runs)
 - provider-target operation route:
   - `POST /v1/provider-targets/operations`
 - product context cutover route:
@@ -1261,22 +1263,26 @@ projections; existing rows and conflicts are reported rather than overwritten.
 The manual `Provider Target Operations` workflow is the supported shared and
 production caller for Phase Two backfill evidence.
 
-Dokploy target setup uses `POST /v1/dokploy-targets/setup`. The route is the
-service-owned path for adopting an existing Dokploy target or creating a new
-application/compose target while immediately writing the matching Dokploy
-target, target-id, and provider-target records. It supports dry-run and apply
-modes, requires `dokploy_target.setup` authz for product/context `launchplane`,
-and requires exact confirmation, an operator reason, and an idempotency key for
-apply. The manual `Dokploy Target Setup` workflow is the supported shared and
-production caller; product repos must not store live target IDs or provider
-fixtures as setup authority. Runtime port is accepted only for `create-compose`
-domain reconciliation with at least one domain. Domain pruning is restricted to
-tracked compose targets and explicit domain hosts; dry-run reports matched
-provider domain ids, while apply deletes only those matching ids and updates the
-tracked target domain list. If a provider create succeeds but the service fails
-before records are written, recover by re-running the workflow with
-`operation=adopt` and the created provider target id, not by creating a second
-target for the same lane.
+Dokploy target setup uses the native FastAPI
+`POST /v1/dokploy-targets/setup` route. The route is the service-owned path for
+adopting an existing Dokploy target or creating a new application/compose target
+while immediately writing the matching Dokploy target, target-id, and
+provider-target records. Its legacy WSGI fallback branch is deleted; direct
+fallback calls fail closed. It supports dry-run and apply modes, requires
+`dokploy_target.setup` authz for product/context `launchplane`, and requires
+exact confirmation, an operator reason, and an idempotency key for apply. Apply
+requests keep the `Idempotency-Key` replay/conflict contract; dry-runs remain
+repeatable and are not stored as idempotency responses. The manual
+`Dokploy Target Setup` workflow is the supported shared and production caller;
+product repos must not store live target IDs or provider fixtures as setup
+authority. Runtime port is accepted only for `create-compose` domain
+reconciliation with at least one domain. Domain pruning is restricted to tracked
+compose targets and explicit domain hosts; dry-run reports matched provider
+domain ids, while apply deletes only those matching ids and updates the tracked
+target domain list. If a provider create succeeds but the service fails before
+records are written, recover by re-running the workflow with `operation=adopt`
+and the created provider target id, not by creating a second target for the same
+lane.
 
 Dokploy target inspect uses the native FastAPI
 `GET /v1/dokploy-targets/inspect` route. The route is a read-only proof surface
